baseline

+++ scene/hexplane.py
@@ -37,8 +37,8 @@ def grid_sample_wrapper(grid: torch.Tensor, coords: torch.Tensor, align_corners:
     B, feature_dim = grid.shape[:2]
     n = coords.shape[-2]
     interp = grid_sampler(
-        grid,  # [B, feature_dim, reso, ...]
-        coords,  # [B, 1, ..., n, grid_dim]
+        grid.float(),  # [B, feature_dim, reso, ...]
+        coords.float(),  # [B, 1, ..., n, grid_dim]
         align_corners=align_corners,
         mode='bilinear', padding_mode='border')
     interp = interp.view(B, feature_dim, n).transpose(-1, -2)  # [B, n, feature_dim]
